+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ## New feature 
 - #473 Pyproject.toml support (@bageljrkhanofemus)
 
+# Improvement
+
+- #479 Add ABC and type hints for TaskHandle and JobSet (@bageljrkhanofemus)
+
 # Release 1.1.1
 
 ## Bug fixes